Create and Update TLS
Can add and update traffic lights to a net file. Can add and update traffic schedules to a net file.
